Got NixOS test VM utility barely working
Impermanence and persistence are really throwing an extra layer of complexity into the mix.
Right now:
- Normal users are non functional.
- A testing user named "swarm", with the password "swarm" is made so login may be done in some fashion.
- The `/persistent/` dir is *not* in the VM for technical reasons.

I suspect `/var/lib/nixos` on is suspect for the issues, as it contains sensitive `uid` and `guid` data.
I theorize that since this is technically a new system—the VM—, the users are entirely new users,
thus attempting to persist/inherit the hosts `uid` and `guid` info into the VM caused some nearly catastrophic failure(s).